Use default class to set result item highlight
Currently the highlight css rules are applied directly to the mark tag,
this causes that utility classes are not applied directly and you need
to define a specific class to override the design.

This change introduces a default_hightlight class to apply the library
colors to the component when no custom class is defined.